Codearena-227: LP > max uint128 are silently lost when staking in RewardsManager - avoid casts from uint256 to uin128 (especially if not safe cast) - change BucketState members to use uint256 to be inline with Pool accounting - add unit test proving that old approach lead to LP lost

+++ src/RewardsManager.sol
@@ -176,8 +176,8 @@ contract RewardsManager is IRewardsManager, ReentrancyGuard {
 
             // update to bucket state
             BucketState storage toBucket = stakeInfo.snapshot[toIndex];
-            toBucket.lpsAtStakeTime  = uint128(positionManager.getLP(tokenId_, toIndex));
-            toBucket.rateAtStakeTime = uint128(IPool(ajnaPool).bucketExchangeRate(toIndex));
+            toBucket.lpsAtStakeTime  = positionManager.getLP(tokenId_, toIndex);
+            toBucket.rateAtStakeTime = IPool(ajnaPool).bucketExchangeRate(toIndex);
             delete stakeInfo.snapshot[fromIndex];
 
             // iterations are bounded by array length (which is itself bounded), preventing overflow / underflow
@@ -233,12 +233,9 @@ contract RewardsManager is IRewardsManager, ReentrancyGuard {
             BucketState storage bucketState = stakeInfo.snapshot[bucketId];
 
             // record the number of lps in bucket at the time of staking
-            bucketState.lpsAtStakeTime = uint128(positionManager.getLP(
-                tokenId_,
-                bucketId
-            ));
+            bucketState.lpsAtStakeTime = positionManager.getLP(tokenId_, bucketId);
             // record the bucket exchange rate at the time of staking
-            bucketState.rateAtStakeTime = uint128(IPool(ajnaPool).bucketExchangeRate(bucketId));
+            bucketState.rateAtStakeTime = IPool(ajnaPool).bucketExchangeRate(bucketId);
 
             // iterations are bounded by array length (which is itself bounded), preventing overflow / underflow
             unchecked { ++i; }
